+++ packages/extensions/pi-todo/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Unbound a session from a todo list that another session deleted or cleared, instead of surfacing an unhandled "Todo list does not exist" error from digests, reminders, and session restore.
 ## [0.4.1] - 2026-08-18
 
 ## [0.4.0] - 2026-08-18